Splitting shallow in two modules
