+++ src/shallow/react.ts
@@ -1,0 +1,13 @@
+import { useRef } from 'react'
+import { shallow } from '../shallow.ts'
+
+export function useShallow<S, U>(selector: (state: S) => U): (state: S) => U {
+  const prev = useRef<U>()
+
+  return (state) => {
+    const next = selector(state)
+    return shallow(prev.current, next)
+      ? (prev.current as U)
+      : (prev.current = next)
+  }
+}
